# docs/TalkRoomsOpenApi.yaml
openapi: 3.0.1
info:
  title: TalkRooms API
  description: 'System to manage message channels and send messages'
  termsOfService: https://swagger.io/terms/
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /invitation/register:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Invitations
      summary: Create a new register invitation
      requestBody:
        description: Details of register invitation
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationInputModelRegister'
      responses:
        '201':
          description: Invitation registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationOutputModelRegister'
        '400':
          description: Email not valid
        '500':
          description: Internal server error
  /invitation/channel:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Invitations
      summary: Create a new channel invitation
      requestBody:
        description: Details of channel invitation
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationInputModelChannel'
      responses:
        '201':
          description: Invitation registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationOutputModelChannel'
        '400':
          description: Email not valid
        '500':
          description: Internal server error
  /invitation/accept/{invitationId}:
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Invitations
      summary: accept invitation
      parameters:
        - name: invitationId
          in: path
          description: Invitation identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          description: User does not correspond to receiver of invitation
        '500':
          description: Internal server error
  /invitation/decline/{invitationId}:
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Invitations
      summary: Decline
      parameters:
        - name: invitationId
          in: path
          description: Invitation identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Invitation accepted
        '400':
          description: User does not correspond to receiver of invitation
        '500':
          description: Internal server error
  /invitation/user/invitations:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Invitations
      summary: Get user invitations
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationList'
        '500':
          description: Internal server error
...
  /messages:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Messages
      summary: Create new message
      description: Create a new message given a sender, a channel, a content and a timestamp
      requestBody:
        description: Message data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageInputModel'
      responses:
        '201':
          description: Message created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOutputModel'
        '400':
          description: Invalid message data
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /messages/{id}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Messages
      summary: Get message by  id
      description: Returns a message
      parameters:
        - name: id
          in: path
          description: message identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Message found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOutputModel'
        '400':
          description: Invalid identifier
        '401':
          description: Unauthorized
        '404':
          description: Message not found
        '500':
          description: Internal server error
  /messages/history/{channelId}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Messages
      summary: Get message by channel id
      description: Returns a list of message
      parameters:
        - name: channelId
          in: path
          description: channel identifier
          required: true
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: number of messages to return
          required: false
          schema:
            type: integer
            example: 20
        - name: skip
          in: query
          description: number of messages to skip
          required: false
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: Messages found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageHistoryOutputModel'
        '400':
          description: Invalid identifier
        '401':
          description: Unauthorized
        '404':
          description: Message not found
        '500':
          description: Internal server error
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserIdentifiers:
      type: object
      description: User details
      required:
        - id
        - username
      properties:
        id:
          type: integer
          description: user id
          example: 1
        username:
          type: string
          description: username
          example: Bob123
...
    ChannelOutputModel:
      type: object
      description: Data that represents a channel
      required:
        - id
        - name
        - creator
        - visibility
      properties:
        id:
          type: integer
          description: channel identifier
          example: 1
        creator:
          $ref: '#/components/schemas/UserIdentifiers'
        name:
          type: string
          description: channel name
          example: DAW
        visibility:
          type: string
          description: channel visibility
          enum: [PUBLIC, PRIVATE]
          example: PUBLIC
...
    InvitationInputModelChannel:
      type: object
      description: Invitation channel input details
      required:
        - receiverId
        - channelId
        - role
      properties:
        receiverId:
          type: integer
          description: receiver identifier
          example: 1
        channelId:
          type: integer
          description: channel identifier
          example: 1
        role:
          type: string
          description: Role of the user in the channel
          enum: [ADMIN, READ_ONLY, READ_WRITE]
          example: READ_ONLY
    InvitationOutputModelChannel:
      type: object
      description: Invitation channel details
      required:
        - id
        - sender
        - receiver
        - channel
        - role
        - timestamp
      properties:
        id:
          type: integer
          description: invitation identifier
          example: 1
        sender:
          $ref: '#/components/schemas/UserIdentifiers'
        receiver:
          $ref: '#/components/schemas/UserIdentifiers'
        channel:
          $ref: '#/components/schemas/ChannelOutputModel'
        role:
          type: string
          description: Role of the user in the channel
          enum: [ADMIN, READ_ONLY, READ_WRITE]
          example: READ_ONLY
        timestamp:
          type: string
          format: date-time
          description: Time the invitation was created
          example: "2024-10-01T12:00:00Z"
    InvitationInputModelRegister:
      type: object
      description: Invitation channel input details
      required:
        - email
        - channelId
        - role
      properties:
        email:
          type: string
          description: receiver email
          example: bob@example.com
        channelId:
          type: integer
          description: channel identifier
          example: 1
        role:
          type: string
          description: Role of the user in the channel
          enum: [ADMIN, READ_ONLY, READ_WRITE]
          example: READ_ONLY
    InvitationOutputModelRegister:
      type: object
      description: Invitation register details
      required:
        - id
        - sender
        - email
        - channel
        - role
        - timestamp
      properties:
        id:
          type: integer
          description: invitation identifier
          example: 1
        sender:
          $ref: '#/components/schemas/UserIdentifiers'
        email:
          type: string
          description: receiver email
          example: bob@example.com
        channel:
          $ref: '#/components/schemas/ChannelOutputModel'
        role:
          type: string
          description: Role of the user in the channel
          enum: [ADMIN, READ_ONLY, READ_WRITE]
          example: READ_ONLY
        timestamp:
          type: string
          format: date-time
          description: Time the invitation was created
          example: "2024-10-01T12:00:00Z"
    InvitationList:
      type: object
      properties:
        nInvitations:
          type: integer
          description: number of results
          example: 20
        channels:
          type: array
          items:
            $ref: '#/components/schemas/InvitationOutputModelChannel'
...
    MessageInputModel:
      type: object
      description: Message details for creation
      required:
        - channelId
        - content
      properties:
        channelId:
          type: integer
          description: channel identifier
        content:
          type: string
          description: Message content
          example: This a message.
    ChannelIdentifiers:
      type: object
      description: Channel details
      required:
        - id
        - name
      properties:
        id:
          type: integer
          description: channel id
          example: 1
        name:
          type: string
          description: Channel name
          example: DAW
    MessageOutputModel:
      type: object
      description: Data that represents a message
      required:
        - msgid
        - sender
        - channel
        - content
        - timestamp
      properties:
        id:
          type: integer
          description: channel identifier
          example: 1
        sender:
          $ref: '#/components/schemas/UserIdentifiers'
        channel:
          $ref: '#/components/schemas/ChannelIdentifiers'
        content:
          type: string
          description: Message content
          example: This is a message.
        timestamp:
          type: string
          format: date-time
          description: Time the message was created
          example: "2024-10-01T12:00:00Z"
    MessageInfoOutputModel:
      type: object
      description: Data that represents a message
      required:
        - msgid
        - sender
        - content
        - timestamp
      properties:
        id:
          type: integer
          description: channel identifier
          example: 1
        sender:
          $ref: '#/components/schemas/UserIdentifiers'
        content:
          type: string
          description: Message content
          example: This is a message.
        timestamp:
          type: string
          format: date-time
          description: Time the message was created
          example: "2024-10-01T12:00:00Z"
    MessageHistoryOutputModel:
      type: object
      description: Data that represents a kist of message
      required:
        - nrOfMessages
        - channel
        - messages
      properties:
        nrOfMessages:
          type: integer
          description: number of results
          example: 1
        channel:
          $ref: '#/components/schemas/ChannelIdentifiers'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/MessageInfoOutputModel'
    User:
      type: object
      description: User details
      required:
        - id
        - username
        - email
      properties:
        id:
          type: integer
          description: user identifier
        username:
          type: string
          description: user username
          example: Bob
        email:
          type: string
          description: user email
          example: bob@example.pt
...
    Channel:
      type: object
      description: Channel details
      required:
        - id
        - name
        - creator
        - visibility
      properties:
        id:
          type: integer
          description: user identifier
        name:
          type: string
          description: Channel name
          example: DAW
        creator:
          $ref: '#/components/schemas/User'
        visibility:
          type: string
          description: channel visibility
          enum: [PUBLIC, PRIVATE]
          example: PUBLIC